学生端周报功能接口第一次测试通过

+++ src/main/java/com/wust_hello/service/student/WeekService.java
@@ -9,6 +9,6 @@ import java.time.LocalDate;
 
 public interface WeekService extends IService<Week> {
     public TotalWeekDto getTotalReport(LocalDate startTime, LocalDate endTime, Integer page, Integer pageSize,String token);
-    public WeekDetailDto getDetails(Integer reportId,String token);
+    public WeekDetailDto getDetails(Long reportId,String token);
     public void updateReport(Week week,String token);
 }
